pamix: use default colors if provided by curses
Terminals do not always have black specified as their default
color, as color code 0 and the default color are actually
distinct values. curse/ncurses provides the
`use_default_colors()` function call to actually enable using
default colors instead of 0 for drawing backgrounds. As this
function is not included in the standard but an extension to the
curses library, we cannot always use it, though.

Check if curses specifies `use_default_colors` and, if so, use it
to have curses draw background colors with the default color
instead of black. Furthermore, initialize color pairs with -1 as
background color in this case, which directs it to use the
default color.