pamix_functions: do not use argument in `pamix_cycle` fns
The `pamix_cycle_prev` and `pamix_cycle_next` functions are used
to cycle through available ports of an entry, either back- or
forwards. But instead of directly calling `cycle_switch` with
either `true` or `false`, which toggle the cycling direction, we
were reading the argument's boolean value and passed that to
`cycle_switch`. As this value was never set, back-cycling does
not work at all.

Fix the issue by directly calling `cycle_switch` with boolean
values instead of using the argument.