pamix: add ability to configure cards
PulseAudio has the concept of audio cards. An audio card may
provide outputs and inputs, depending on its configuration. Many
cards also allow to change their configuration at runtime to e.g.
switch between different outputs of the card. Depending on which
configuration was selected for a card, it then provides a
different set of output- and input-devices. This may for example
be used to switch a card between outputting via analog and HDMI
or to enable duplex.

This commit adds the ability to configure cards by switching
between their profiles, where each profile select a certain
configuration of outputs and inputs.