pamix: conditionally draw an entry's meter
Right now, all entries have an output or input volume. As such,
it does make sense to unconditionally draw an entry's meter, as
every device has one. But as we want to introduce a new entry
type for sound cards, which do not have a meter which would be
displayed, we want to make drawing meters conditional on an entry
actually having one.

Introduce a new variable `m_Meter` to the `Entry` structure which
states if the entry has a meter. Use the boolean to only draw
meters if the entry has one.